Organize comparers, exposing a couple, and add to changelog.
